+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- disable native mobile app viewport screenshotting (until check region functionality is also supported)
 
 ## 12.4.0 - 2020/10/13